+++ packages/client/hmi-client/src/components/widgets/VegaChart.vue
@@ -26,31 +26,11 @@
 
 <script setup lang="ts">
 import _ from 'lodash';
-import { format } from 'd3';
 import embed, { Config, Result, VisualizationSpec } from 'vega-embed';
 import Button from 'primevue/button';
 import Dialog from 'primevue/dialog';
-import { countDigits, fixPrecisionError } from '@/utils/number';
 import { ref, watch, toRaw, isRef, isReactive, isProxy, computed, h, render, onUnmounted } from 'vue';
-
-const NUMBER_FORMAT = '.3~s';
-
-// Define the custom expression functions that can be registered and used in the Vega charts
-const expressionFunctions = {
-	// chartNumberFormatter is a custom number format that will display numbers in a more readable format
-	chartNumberFormatter: (value: number) => {
-		const correctedValue = fixPrecisionError(value);
-		if (value > -1 && value < 1) {
-			return countDigits(correctedValue) > 6 ? correctedValue.toExponential(3) : correctedValue.toString();
-		}
-		return format(NUMBER_FORMAT)(correctedValue);
-	},
-	// Just show full value in tooltip
-	tooltipFormatter: (value) => {
-		if (value === undefined) return 'N/A';
-		return fixPrecisionError(value);
-	}
-};
+import { expressionFunctions } from '@/services/charts';
 
 // This config is default for all charts, but can be overridden by individual chart spec
 const defaultChartConfig: Partial<Config> = {
